+++ io/CHANGES.md
@@ -1,4 +1,10 @@
-# unreleased
+# unreleased 0.3.0
+## Change
+- rename `http3` feature to `quic`
+- rename `H3ServerConfig` to `QuicConfig`
+- rename `Udp` prefixed types to `Quic` prefixed
+- `AsyncIo::ready` and `AsyncIoDyn::ready` receive `&mut self`. Allowing more arbitrary types implement these traits.
+- `AsyncIo::poll_ready` and `AsyncIoDyn::poll_ready` receive `&mut self`. For the same reason as above change.
 
 # 0.2.1
 ## Add